I changed the standard display to show that the plant is happy when there are now warnings Such as: All is well The plant is happy And then a happy emoji afterwards.
Then I changed the warning to display when
there is 1 or more warnings in the system.

I've also changed the order of which the functions
are called, so that the sensors are
called first to count the warnings before
we display them, play a sad tune and turn
the light red.

Within the warning I've placed an if-statement
for the light sensor's warning of too
much light.

I expect the warnings for the other sensors to be right after it.
But don't write 'else if' because then
only one warning is displayed, even if all
the sensors have a warning.